add regulator kmods and set them as device specific packages
The GL-inet ax1800 (and probably the axt1800 as well) is missing the usual
MP5496 SOC regulator. Instead it's just using a fixed SOC voltage.

To reflect that different kernel config, 2 kmod's
have been created. One for the MP5496 and one for the fixed regulator.

Also including a dts change for the ax1800 for the required
opp and regulator adaption.

Signed-off-by: Dirk Buchwalder <buchwalder@posteo.de>